Moved SpecDoc with losses to own class
